+++ src/drv/ds18b20_drv.h
@@ -8,6 +8,6 @@ unsigned char ds18B20_check(void);
 unsigned char ds18b20_init(void) ;
 void write_byte(unsigned char val);
 unsigned char read_byte(void);
-short Read_Temp(void);
+float Read_Temp(void);
 
 #endif
